RHOL-11: Rholang: Move RholangOnlyDispatcher to test package.
